fix: BROS-66: Fix Syncable in Interactive View All
Another try to fix this.
Synced tags create SyncManager instance for every group in every annotation.
With DEV-3391 FF OFF we had only one config tree and all tags in group were refering
to the same manager.
With DEV-3391 FF ON we have separate config trees so we should distinguish managers.

BUT
History items also trigger `afterCreate()` from `SyncableMixin`, but we don't have
annotation there.
One of the ways to fix this is to modify AnnotationMixin to retrieve HistoryItem in
such cases.