fix: BROS-78: Fix "No more annotations" with Taxonomy in Preview
If you add Taxonomy to your config in Preview or edit config with Taxonomy in it,
most likely you'll see "No more annotations" info block instead of preview.
This is happening because of wrong usage of lazy reference in SharedChoiceMixin in
`afterCreate()` - it should attach store to a tag when reference is missing, but
it fails upon access. Using recommended `tryReference()` makes it safe and fixes
the original issue.