fix: BROS-325: Don't show hidden per-regions in Info panel
We store conditional per-regions in results even if the condition is not met
the moment for better UX, so user can switch between different options and
see previous choices for all of them.
But hidden choices won't be submitted.
And they also should not be visible in Info panel to not confuse the user.
This fix is dedicated to hide them.